Improvements to generated code for generic property types.
Note that method call 'equalTo(x)' where x is a Box<? extends Widget>)
returns a matcher of type
Matcher<? extends Box<? extends Widget>> by default. This is
undesirable, as it can't then be assigned to a PropertyMatcher that
accepts Matcher<? super Box<? extends Widget>>. This problem can be
avoided by explicitly stating the type, as in:
CoreMatcher.<Box<? extends Widget>>equalTo(x).

This is only necessary when the type of x includes a wildcard at the top
level. If x has type <Box<Box<? extends Widget>> then the return type
is Matcher<Box<Box<? extends Widget>>>, which is fine.