Fix failing tests by enforcing a defined order on the property list returned by PropertyLocator.
Tests failed because the order of properties in the generated file did not
match the order in the expected result. Enforcing an ordering on the properties
should prevent future failures.

The tests also failed to locate the test data due to a case difference in
the specified directory name.

Refactored other aspects of Property and PropertyLocator.